+++ tests/test_cons.py
@@ -133,7 +133,6 @@ def test_cons_indicator_fail():
     m.setSolVal(sol, binvar, 0)
     assert m.checkSol(sol)  # solution should be feasible
 
-
 def test_addConsCardinality():
     m = Model()
     x = {}
@@ -144,6 +143,22 @@ def test_addConsCardinality():
     m.optimize()
 
     assert m.isEQ(m.getVal(quicksum(x[i] for i in range(5))), 3)
+
+def test_getOrigConss():
+    m = Model()
+    x = m.addVar("x", lb=0, ub=2, obj=-1)
+    y = m.addVar("y", lb=0, ub=4, obj=0)
+    z = m.addVar("z", lb=0, ub=5, obj=2)
+    m.addCons(x <= y + z)
+    m.addCons(x <= z + 100)
+    m.addCons(y >= -100)
+    m.addCons(x + y <= 1000)
+    m.addCons(2* x + 2 * y <= 1000)
+    m.addCons(x + y + z <= 7)
+    m.optimize()
+    assert len(m.getConss(transformed=False)) == m.getNConss(transformed=False)
+    assert m.getNConss(transformed=False) == 6
+    assert m.getNConss(transformed=True) < m.getNConss(transformed=False)
 
 
 def test_printCons():